+++ src/mol-repr/volume/util.ts
@@ -1,15 +1,17 @@
 /**
- * Copyright (c) 2020 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2020-2022 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
  */
 
 import { Volume } from '../../mol-model/volume';
 import { Loci } from '../../mol-model/loci';
-import { Interval, OrderedSet } from '../../mol-data/int';
+import { Interval, OrderedSet, SortedArray } from '../../mol-data/int';
 import { equalEps } from '../../mol-math/linear-algebra/3d/common';
 import { Vec3 } from '../../mol-math/linear-algebra/3d/vec3';
 import { packIntToRGBArray } from '../../mol-util/number-packing';
+import { SetUtils } from '../../mol-util/set';
+import { Box3D } from '../../mol-math/geometry';
 
 // avoiding namespace lookup improved performance in Chrome (Aug 2020)
 const v3set = Vec3.set;
@@ -19,18 +21,17 @@ const v3addScalar = Vec3.addScalar;
 const v3scale = Vec3.scale;
 const v3toArray = Vec3.toArray;
 
-export function eachVolumeLoci(loci: Loci, volume: Volume, isoValue: Volume.IsoValue | undefined, apply: (interval: Interval) => boolean) {
+export function eachVolumeLoci(loci: Loci, volume: Volume, props: { isoValue?: Volume.IsoValue, segments?: SortedArray } | undefined, apply: (interval: Interval) => boolean) {
     let changed = false;
     if (Volume.isLoci(loci)) {
         if (!Volume.areEquivalent(loci.volume, volume)) return false;
         if (apply(Interval.ofLength(volume.grid.cells.data.length))) changed = true;
     } else if (Volume.Isosurface.isLoci(loci)) {
         if (!Volume.areEquivalent(loci.volume, volume)) return false;
-        if (isoValue) {
-            if (!Volume.IsoValue.areSame(loci.isoValue, isoValue, volume.grid.stats)) return false;
+        if (props?.isoValue) {
+            if (!Volume.IsoValue.areSame(loci.isoValue, props.isoValue, volume.grid.stats)) return false;
             if (apply(Interval.ofLength(volume.grid.cells.data.length))) changed = true;
         } else {
-            // TODO find a cheaper way?
             const { stats, cells: { data } } = volume.grid;
             const eps = stats.sigma;
             const v = Volume.IsoValue.toAbsolute(loci.isoValue, stats).absoluteValue;
@@ -48,6 +49,27 @@ export function eachVolumeLoci(loci: Loci, volume: Volume, isoValue: Volume.IsoV
             OrderedSet.forEach(loci.indices, v => {
                 if (apply(Interval.ofSingleton(v))) changed = true;
             });
+        }
+    } else if (Volume.Segment.isLoci(loci)) {
+        if (!Volume.areEquivalent(loci.volume, volume)) return false;
+        if (props?.segments) {
+            if (!SortedArray.areIntersecting(loci.segments, props.segments)) return false;
+            if (apply(Interval.ofLength(volume.grid.cells.data.length))) changed = true;
+        } else {
+            const segmentation = Volume.Segmentation.get(volume);
+            if (segmentation) {
+                const set = new Set<number>();
+                for (let i = 0, il = loci.segments.length; i < il; ++i) {
+                    SetUtils.add(set, segmentation.segments.get(loci.segments[i])!);
+                }
+                const s = Array.from(set.values());
+                const d = volume.grid.cells.data;
+                for (let i = 0, il = d.length; i < il; ++i) {
+                    if (s.includes(d[i])) {
+                        if (apply(Interval.ofSingleton(i))) changed = true;
+                    }
+                }
+            }
         }
     }
     return changed;
@@ -179,4 +201,49 @@ export function createVolumeTexture3d(volume: Volume) {
     }
 
     return textureVolume;
+}
+
+export function createSegmentTexture2d(volume: Volume, set: number[], bbox: Box3D, padding = 0) {
+    const data = volume.grid.cells.data;
+    const dim = Box3D.size(Vec3(), bbox);
+    const o = volume.grid.cells.space.dataOffset;
+    const { width, height } = getVolumeTexture2dLayout(dim, padding);
+
+    const itemSize = 1;
+    const array = new Uint8Array(width * height * itemSize);
+    const textureImage = { array, width, height };
+
+    const [xn, yn, zn] = dim;
+    const xn1 = xn - 1;
+    const yn1 = yn - 1;
+    const zn1 = zn - 1;
+
+    const xnp = xn + padding;
+    const ynp = yn + padding;
+
+    const [minx, miny, minz] = bbox.min;
+    const [maxx, maxy, maxz] = bbox.max;
+
+    for (let z = 0; z < zn; ++z) {
+        for (let y = 0; y < yn; ++y) {
+            for (let x = 0; x < xn; ++x) {
+                const column = Math.floor(((z * xnp) % width) / xnp);
+                const row = Math.floor((z * xnp) / width);
+                const px = column * xnp + x;
+                const index = itemSize * ((row * ynp * width) + (y * width) + px);
+
+                const v0 = set.includes(data[o(x + minx, y + miny, z + minz)]) ? 255 : 0;
+                const xp = set.includes(data[o(Math.min(xn1 + maxx, x + 1 + minx), y + miny, z + minz)]) ? 255 : 0;
+                const xn = set.includes(data[o(Math.max(0, x - 1 + minx), y + miny, z + minz)]) ? 255 : 0;
+                const yp = set.includes(data[o(x + minx, Math.min(yn1 + maxy, y + 1 + miny), z + minz)]) ? 255 : 0;
+                const yn = set.includes(data[o(x + minx, Math.max(0, y - 1 + miny), z + minz)]) ? 255 : 0;
+                const zp = set.includes(data[o(x + minx, y + miny, Math.min(zn1 + maxz, z + 1 + minz))]) ? 255 : 0;
+                const zn = set.includes(data[o(x + minx, y + miny, Math.max(0, z - 1 + minz))]) ? 255 : 0;
+
+                array[index] = Math.round((v0 + v0 + xp + xn + yp + yn + zp + zn) / 8);
+            }
+        }
+    }
+
+    return textureImage;
 }